Move getClassFromId function into adminclass
It's the only place it was ever used.
